+++ lib/fabricio/cli/cli.rb
@@ -2,14 +2,14 @@ require 'thor'
 require 'fabricio'
 require 'fileutils'
 require 'yaml'
+require_relative 'organization'
+require_relative 'cli_helper'
 
 module Fabricio
   class CLI < Thor
 
-    # Constants
-    CREDENTIAL_DIRECTORY_PATH = "#{Dir.home}/.fabricio"
-    CREDENTIAL_FILE_PATH = "#{CREDENTIAL_DIRECTORY_PATH}/.credential"
-    FABRIC_GRAPHQL_API_URL = 'https://api-dash.fabric.io/graphql'
+    desc "organization", "..."
+    subcommand "organization", Organization
 
     desc "credential", "Setup credential"
     def credential
@@ -30,17 +30,6 @@ module Fabricio
       end
 
       say("Complete!")
-    end
-
-    desc "organization", "Obtain organization"
-    option :app_id => :required, :type => :string
-    option :short, :type => :boolean
-    def organization
-      if options[:short]
-        say(client.organization.get.pretty_print)
-      else
-        say("#{client.organization.get.to_s}")
-      end
     end
 
     desc "apps", "Obtain all app"
@@ -88,45 +77,6 @@ module Fabricio
       else
         say("#{client.build.get(app_id, version, build_number).to_s}")
       end
-    end
-
-    private
-    def client
-      email = ""
-      password = ""
-      if File.file?(CREDENTIAL_FILE_PATH)
-        credential = YAML.load_file(CREDENTIAL_FILE_PATH)
-        email = credential['email']
-        password = credential['password']
-      else
-        ask_credential
-      end
-
-      client = Fabricio::Client.new do |config|
-        config.username = email
-        config.password = password
-      end
-    end
-
-    def create_credential_file(credential)
-      FileUtils.mkdir_p(CREDENTIAL_DIRECTORY_PATH)
-      credential_hash = {
-        "email" => credential.email,
-        "password" => credential.password
-      }
-      File.open(CREDENTIAL_FILE_PATH,'w') do |f|
-          f.write credential_hash.to_yaml
-      end
-      say("Your credential in #{CREDENTIAL_FILE_PATH}")
-    end
-
-    def ask_credential
-      say("We have to know you're email from fabric account")
-      email = ask("email: ")
-      say("Now we want your password. Do not be afraid, it is stored locally")
-      password = ask("password: ", :echo => false)
-      say("")
-      Fabricio::Model::Credential.new(email, password)
     end
 
   end
